feat: Add CreateMeasurementForm component for user measurement input
